Increase timing precision to centiseconds
The hundredths digit of a second is the standard for cubing
competitions. Speedstacking uses milliseconds.

To reflect this, the timing system had to be changed, as incrementing
would rapidly cause drift from real timing. Singh's original 1 s gap
meant the drift wasn't too noticeable over short durations, but with
updates a hundred times a second, it adds up much more quickly.

Also, the UI was edited so seconds are rounded, since floating-point
numbers have a tendency to inflate with insignificant digits due to
their specification.